OpenAI-NLPtoSQL/requirements.txt
openai
pandas
sqlalchemy
python-dotenv
mysql-connector-python
flask
flask-cors
flask_login
flask_bcrypt